+++ di-monad/default.nix
@@ -1,12 +1,13 @@
 { mkDerivation, base, containers, di-core, exceptions, lib, mtl
-, pipes, stm, transformers
+, pipes, stm, streaming, transformers, unliftio-core
 }:
 mkDerivation {
   pname = "di-monad";
-  version = "1.3.1";
+  version = "1.3.2";
   src = ./.;
   libraryHaskellDepends = [
-    base containers di-core exceptions mtl pipes stm transformers
+    base containers di-core exceptions mtl pipes stm streaming
+    transformers unliftio-core
   ];
   homepage = "https://github.com/k0001/di";
   description = "mtl flavoured typeful hierarchical structured logging for di-core";